Refactor handlers to use ResponseResult and add auth checks

+++ Core/CleanArchitectureTemplate.Application/Features/Products/Commands/DeleteProduct/DeleteProductCommandHandler.cs
@@ -1,11 +1,12 @@
 ﻿using AutoMapper;
 using CleanArchitectureTemplate.Application.Abstractions.Repositories.Files;
 using CleanArchitectureTemplate.Application.Abstractions.Repositories.Shopping;
+using CleanArchitectureTemplate.Application.Abstractions.Services;
 using CleanArchitectureTemplate.Application.Common.Responses;
 using CleanArchitectureTemplate.Application.Dtos.Shopping;
+using CleanArchitectureTemplate.Application.Exceptions;
 using CleanArchitectureTemplate.Application.Features.ProductImageFiles.Commands.DeleteProductImagesByProductId;
 using CleanArchitectureTemplate.Domain.Entities.Shopping;
-using CleanArchitectureTemplate.Application.Exceptions;
 using MediatR;
 
 namespace CleanArchitectureTemplate.Application.Features.Products.Commands.DeleteProduct;
@@ -19,13 +20,15 @@ public class DeleteProductCommandHandler : IRequestHandler<DeleteProductCommandR
     private readonly IMediator _mediator;
     private readonly IProductReadRepository _productReadRepository;
     private readonly IProductWriteRepository _productWriteRepository;
+    private readonly IUserContextService _userContextService;
 
-    public DeleteProductCommandHandler(IMapper mapper, IMediator mediator, IProductReadRepository productReadRepository, IProductWriteRepository productWriteRepository, IProductImageFileWriteRepository productImageFileWriteRepository)
+    public DeleteProductCommandHandler(IMapper mapper, IMediator mediator, IProductReadRepository productReadRepository, IProductWriteRepository productWriteRepository, IProductImageFileWriteRepository productImageFileWriteRepository, IUserContextService userContextService)
     {
         _mapper = mapper;
         _mediator = mediator;
         _productReadRepository = productReadRepository;
         _productWriteRepository = productWriteRepository;
+        _userContextService = userContextService;
     }
 
     public async Task<SingleResponse<ProductDto?>> Handle(DeleteProductCommandRequest request, CancellationToken cancellationToken)
@@ -34,6 +37,9 @@ public class DeleteProductCommandHandler : IRequestHandler<DeleteProductCommandR
 
         var product = await _productReadRepository.GetByIdAsync(request.Id)
             ?? throw new NotFoundException(nameof(Product), request.Id);
+
+        if (!_userContextService.IsAdminOrSelf(product.StoreId))
+            throw new ForbiddenException();
 
         var deleteProductImagesCommand = new DeleteProductImagesByProductIdCommandRequest(product!.Id);
         await _mediator.Send(deleteProductImagesCommand, cancellationToken);
